Select subject URIs in pages and load documents in batches
The subject select query was read in a lazy-seq which meant (on the observation
pipeline where there are _a lot_ of resources) the connection could be closed by
drafter before all the results were read (leading to `TruncatedChunkException`s).

We now execute the select query in pages with limit/ offset. This obviates the need
to partition the results before extracting resources descriptions.

This patch also batches the loading into ES - we can extract more resources at a
time from drafter than we can load into ES causing the inserts to be rejected
with "Too Many Requests" - thus the doc batch size is smaller than the subject page size.