Replace codelist search with code search
`datasets/for-facets` is the workhorse function. This performs the structural search query and gathers the rest of the data needed to explain the matches. The structural search is done over observations, with the results being returned in terms of datasets.

The feature is integration tested in `ook.search.elastic-test` with `ook.search.db/get-datasets-for-facets`.

The `ook.search.elastic.datasets-test` unit test `datasets/datasets-from-observation-hits` and `datasets/explain-match` without needing the database.

The `facets` namespace no longer does any dataset manipulation/ filtering. The `apply-facets` and `apply-selections` functions are replaced with `dimension-selections` and `get-facets-for-selections`.

The `components` namespace now includes `get-codelists` and `codelist-to-dimensions-lookup`.

The `codes` namespace has a `get-codes` function to look-up codes by URI ("terms"). This will need distinguishing from text queries ("match") once those are reintroduced.