Add 'used' property to code index
This property describes whether the code is in use. It ought ultimately be considered mutable given
that this status can change as new observations are loaded.

Sadly we're writing this to the index as a string - it appears that the jsonld library doesn't coerce types to native json types.

We can review this later but for now it should be easy enough to parse the string representation (indeed ES is indifferent).