Give drafter client dummy AUTH0 creds
We had previously configured drafter-client with a real AUTH0 user and password.

It turns out these aren't used while we're only downloading RDF from the live endpoint.

This patch comments out the real creds (so the wiring is still there if we need to use
a draft endpoint in future).