Page observation pipeline subject by graph
The subject-cache for the observation pipeline has become too big.

We need to handle 30m subjects and more.

Downloading these with limit/offset was impossible as we can't sort/ retreive later pages quickly enough.

Even a single download exceeded the 15 minute stardog timeout and caused drafter to hang in GC (presumably because stasher would hold the whole 3G+ result in memory).

This patch introduces an optional `graph-query` parameter that's applied on the `observation-pipeline`. We use this to page the subject-query by graph (then into page-size partitions).

Hopefully this fixes #103.