Revert "Attempt smaller page size for codes-used pipeline"
This reverts commit 8c8e44512cbc25e53183ba0b429cff883f8014e1.

Although only a config change, the commit we're reverting would've required the
fixtures cassette to be refreshed. The new data is quite different and doesn't
look quite right. Rather than work through this now I'm reverting the commit.

The reverted commit was only intended to reduce the page size on the codes-used
pipeline. Now that write-activity on the beta site has slowed down it seems the etl
run is succeeding with the old config anyway!