Results snippet takes matches from code search not by aggregating observations
The dimension-rollup aggregation in the observation search query takes too long,
as much as 15 seconds for "imports of cars from Germany".

This is used to identify matches to return in the rich snippet. We can instead use
the original list of matching codes that was used to filter observations in the
first place. This is of course much faster than aggregating (potentially millions)
of observations to find the distinct values per component.

This will lead to false positives - i.e. codes that matched the search and are
present in the dataset's codelist but for which no observations exist. The false
positives won't apply to the matching-observation count (which still uses the
observations index). They'll also be ignored by Muttnik's cube-explorer.